+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to the code converter will be documented here.
 * Convert types in ternary expressions (#363)
 * Support for converting dot net standard VB projects (#398)
 * Avoid compilation error for duplicate cases (#374)
+* Correctly handle type promoted module symbols (#375)
 
 ### C# -> VB
 * Convert property accessors with visiblity modifiers (#92)